+++ oci/load_balancer_load_balancer_test.go
@@ -35,13 +35,13 @@ var (
 		"compartment_id": Representation{repType: Required, create: `${var.compartment_id}`},
 		"display_name":   Representation{repType: Required, create: `example_load_balancer`, update: `displayName2`},
 		"shape":          Representation{repType: Required, create: `100Mbps`},
-		"subnet_ids":     Representation{repType: Required, create: []string{"${oci_core_subnet.lb_test_subnet_1.id}", "${oci_core_subnet.lb_test_subnet_2.id}"}},
+		"subnet_ids":     Representation{repType: Required, create: []string{`${oci_core_subnet.lb_test_subnet_1.id}`, `${oci_core_subnet.lb_test_subnet_2.id}`}},
 		"defined_tags":   Representation{repType: Optional, create: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "value")}`, update: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "updatedValue")}`},
 		"freeform_tags":  Representation{repType: Optional, create: map[string]string{"Department": "Finance"}, update: map[string]string{"Department": "Accounting"}},
 		"is_private":     Representation{repType: Optional, create: `false`},
 	}
 
-	LoadBalancerSubnetDependencies = `
+	LoadBalancerSubnetDependencies = AvailabilityDomainConfig + `
 	data "oci_load_balancer_shapes" "t" {
 		compartment_id = "${var.compartment_id}"
 	}
@@ -52,7 +52,7 @@ var (
 
 	resource "oci_core_subnet" "lb_test_subnet_1" {
 		#Required
-		availability_domain = "${lookup(data.oci_identity_availability_domains.ADs.availability_domains[0],"name")}"
+		availability_domain = "${data.oci_identity_availability_domains.test_availability_domains.availability_domains.0.name}"
 		cidr_block = "10.0.0.0/24"
 		compartment_id = "${var.compartment_id}"
 		vcn_id = "${oci_core_vcn.test_vcn.id}"
@@ -62,7 +62,7 @@ var (
 	
 	resource "oci_core_subnet" "lb_test_subnet_2" {
 		#Required
-		availability_domain = "${lookup(data.oci_identity_availability_domains.ADs.availability_domains[1],"name")}"
+		availability_domain = "${data.oci_identity_availability_domains.test_availability_domains.availability_domains.1.name}"
 		cidr_block = "10.0.1.0/24"
 		compartment_id = "${var.compartment_id}"
 		vcn_id = "${oci_core_vcn.test_vcn.id}"