+++ CHANGELOG.md
@@ -1,4 +1,31 @@
-## 3.6.1 (Unreleased)
+## 3.9.1 (Unreleased)
+## 3.9.0 (December 04, 2018)
+
+### Added
+- Support for the Instance Pools & Instance Configurations
+- Support for the Block Volume cross-region backups
+- Support for 'approximate_count' and 'approximate_size' for bucket resource
+
+## 3.8.0 (November 28, 2018)
+
+### Added
+- Support VCN Transit
+
+## 3.7.0 (November 14, 2018)
+
+### Added
+- New parameter `is_hydrated` in `oci_core_volume_groups` resource and data source
+- Support for public IP prefixes (CIDRs) up to 31
+- Support for tagging in `oci_file_storage_file_system`, `oci_file_storage_mount_target`, and `oci_file_storage_snapshot`
+
+### Changed
+- Make `route_table_id`, `dhcp_options_id` in `oci_core_subnet` updatable
+- Make `security_list_ids` in `oci_core_subnet` optional and updatable
+
+### Deprecated
+- Volumes: The `backup_policy_id` attribute is now deprecated. Backup policy should be assigned through `volume_backup_policy_assignments` resource instead.
+- BootVolumes: The `backup_policy_id` attribute is now deprecated. Backup policy should be assigned through `volume_backup_policy_assignments` resource instead.
+
 ## 3.6.0 (November 01, 2018)
 
 ### Added